% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_info}
\alias{df_info}
\title{Database seedling and rosettes}
\format{
A dataset with nine columns:
\describe{
  \item{year}{A string, it's red.}
  \item{site}{A string, it's green. Street talk for money.}
  \item{blue}{A string, really didn't see that coming. It's blue this time!}
}
}
\source{
\url{http://www.themindofsebastian.com}
}
\usage{
df_info
}
\description{
A dataset containing nine columns describing count, survival and death
of seedling and rosettes from one year of census to the next
}
\keyword{datasets}
