% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosettes.R
\name{info_transform}
\alias{info_transform}
\title{create rosettes cleaned database

create rosettes cleaned database}
\usage{
info_transform(df_info, year_dependency = FALSE)
}
\arguments{
\item{df_info}{initial database}

\item{year_dependency}{allows for plants generation taking into account previous year}
}
\value{
cleaned database
}
\description{
transforms the rosettes info data frame into the data cleaned format
}
