% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedling.R
\name{info_transform_seedling}
\alias{info_transform_seedling}
\title{create seedling cleaned database

create seedling cleaned database}
\usage{
info_transform_seedling(df_info, year_dependency = FALSE)
}
\arguments{
\item{df_info}{initial database}

\item{year_dependency}{allows for plants generation taking into account previous year}
}
\value{
cleaned database that i will use .. ???
}
\description{
transforms the seedling info data frame into the data cleaned format
}
