% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedling_death.R
\name{random_death}
\alias{random_death}
\title{random_death

create a function that randomly selct the seedling who has to die}
\usage{
random_death(df, num_deaths)
}
\arguments{
\item{df}{dataframe}

\item{num_deaths}{}
}
\value{

}
\description{
kills plt for the seedling info data frame
}
\examples{
info_transform_seedling()
}
